Duplicate instead of sharing
